+++ database/migrations/2023_11_26_062014_create_sales_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->date('date');
             $table->string('reference');
-            $table->integer('total_amount');
-            $table->integer('paid_amount');
-            $table->integer('due_amount');
+            $table->bigInteger('total_amount');
+            $table->bigInteger('paid_amount');
+            $table->bigInteger('due_amount');
             $table->string('status');
             $table->string('payment_status');
             $table->string('payment_method');
